Add DiscMesh class to store meshes for DiscRenderer.
