+++ Assets/Teatro/Disc/DiscMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02f07a361d9f2e0448b1e28750646c06
+timeCreated: 1451315803
+licenseType: Pro
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
