+++ Assets/Teatro/Disc/DiscMesh.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 8ddd33a796ecb0846ae35b516e9dc6e9
+timeCreated: 1451313176
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
