+++ Assets/Teatro/Disc/Editor/DiscMeshEditor.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c1fd8773e0204d0478c27abe725e066a
+timeCreated: 1451313908
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
